refactoring and annotating/improving security
